requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
